refactor: List 컴포넌트 파일명 List로 통일
- Container였던 흔적 삭제
